Don't delete old releases from the website
It's convenient for old releases to be available indefinitely so that
deployment scripts can retrieve a particular qualified version.  This
won't affect the disk space requirements of the repository, since the
old packages will be in the Git history anyhow.

Move packages to a subdirectory of the site root to avoid causing
permanent clutter.