Fix inherited error in clause 3 of LICENSE file
The original AJAX Control Toolkit license prohibited the name of the
CodePlex Foundation from being used to endorse/promote derived products.
We need to preserve this requirement.  This was already done in the .js
header blocks, and should have been done in LICENSE.txt in a7c96c56 but
was missed.

Original discussion in #10.  Permission for the license change was given
by all existing contributors in #58.