IE 10 not reading DZI file correctly
IE 10 is treating the data coming back from the JSONP request as a
string and not as XML. I have confirmed this issue is happening on
numerous IE10 machines but have not seen it on any other browser.

The change simply checks the type of the data variable and if it is a
string it parses the string as XML and updates the data object.