Make tileSources option smarter about detecting when a json string or xml string has been passed in
The tileSources option was only using a test looking for the presence of a "{", "[", or "<" character to determine if a json string or xml string was passed in.  It is possible for a url to contain one of these characters as well which would break using the url as a tileSources parameter.  The following is an example of a breaking url:
http://myurl.org/{25fb14f0-a839-4c4d-8c97-dd1d67b2cb35}/MyImage.xml

This patch resolves this issue.