Improved testing data for IIIFv3
This commit adds a new set of tests for IIIFv3 Image API info
to the `formats` tests. These tests mirror the tests done for the v2
handler, but with updated 'info.json' files.

Image directories also needed to be changed to include the full width/height
parameters.

Also changed was moving the `iiif` test block up below formats, and adding a QUnit
module name. This allowed the tests to show up in the QUnit browser test.

All of the new tests were checked and pass.