Add support for IIP (Internet Imaging Protocol) API
This pull request adds a new tilesource for the IIP API which is compatible with IIP servers such as IIPImage. The tilesource supports a number of image processing transforms.
Configuration requires only the specification of the API end-point `iipsrv` and the image name `image`. The optional `format` parameter can be used to request tiles in image formats other than JPEG, such as WebP or AVIF. The `transform` parameter can also be optionally used to specify an object containing a set of image processing transforms.

Example: tileSources: {
            iipsrv:    "/fcgi-bin/iipsrv.fcgi",
            image:     "test.tif",
            format:    "webp",
            transform: {
                         gamma: 1.5,
                         invert: true
                       }
          }